chore: moved ad rendering into the article
the markup component is meant to only deal with text formatting by default.
