+++ packages/markup/__tests__/markup.native.test.js
@@ -3,5 +3,4 @@ import { Text, View } from "react-native";
 import runTests from "./test-helper";
 import { renderTree, renderTrees } from "../markup";
 
-jest.mock("WebView", () => "WebView");
 describe("Markup Native", runTests(renderTree, renderTrees, Text, View));
