+++ packages/markup/__tests__/test-helper.js
@@ -5,8 +5,6 @@ import renderer from "react-test-renderer";
 
 const singleParagraph = require("../fixtures/single-paragraph.json").fixture;
 const multiParagraph = require("../fixtures/multi-paragraph.json").fixture;
-const multiParagraphWithAds = require("../fixtures/multi-paragraph-with-ads.json")
-  .fixture;
 const anchor = require("../fixtures/anchor.json").fixture;
 const bold = require("../fixtures/bold.json").fixture;
 const italic = require("../fixtures/italic.json").fixture;
@@ -32,16 +30,6 @@ export default (
   it("renders multiple paragraphs", () => {
     const output = renderer
       .create(<BlockComponent>{renderTrees(multiParagraph)}</BlockComponent>)
-      .toJSON();
-
-    expect(output).toMatchSnapshot();
-  });
-
-  it("renders multiple paragraphs with ads", () => {
-    const output = renderer
-      .create(
-        <BlockComponent>{renderTrees(multiParagraphWithAds)}</BlockComponent>
-      )
       .toJSON();
 
     expect(output).toMatchSnapshot();
