--- packages/markup/fixtures/multi-paragraph-with-ads.json
@@ -1,135 +1,0 @@
-{
-  "fixture": [
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph one"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "ad",
-      "attributes": {},
-      "children": []
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "name": "paragraph",
-      "attributes": {},
-      "children": [
-        {
-          "name": "text",
-          "children": [
-            {
-              "text": "This is the text for paragraph two"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
