+++ packages/markup/package.json
@@ -52,7 +52,6 @@
     "webpack": "3.3.0"
   },
   "dependencies": {
-    "@times-components/ad": "0.3.1",
     "prop-types": "15.5.10",
     "react-native-web": "0.0.119"
   },
